start implementing export changes

+++ src/main/java/de/tum/in/www1/artemis/service/SubmissionExportService.java
@@ -5,7 +5,11 @@ import java.io.IOException;
 import java.nio.file.Path;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Comparator;
+import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 import javax.annotation.Nullable;
@@ -58,8 +62,12 @@ public abstract class SubmissionExportService {
      * @return the zipped file with the exported submissions
      */
     public File exportStudentSubmissionsElseThrow(Long exerciseId, SubmissionExportOptionsDTO submissionExportOptions) {
-        return exportStudentSubmissions(exerciseId, submissionExportOptions)
-                .orElseThrow(() -> new BadRequestAlertException("Failed to export student submissions.", "SubmissionExport", "noSubmissions"));
+        var zippedSubmissionsPaths = exportStudentSubmissions(exerciseId, submissionExportOptions);
+        if (zippedSubmissionsPaths.isEmpty()) {
+            throw new BadRequestAlertException("Failed to export student submissions.", "SubmissionExport", "noSubmissions");
+        }
+        return zippedSubmissionsPaths.get(0).toFile();
+
     }
 
     /**
@@ -70,20 +78,20 @@ public abstract class SubmissionExportService {
      * @param submissionExportOptions the options for the export
      * @return the zipped file with the exported submissions
      */
-    public Optional<File> exportStudentSubmissions(Long exerciseId, SubmissionExportOptionsDTO submissionExportOptions) {
+    public List<Path> exportStudentSubmissions(Long exerciseId, SubmissionExportOptionsDTO submissionExportOptions) {
         Path outputDir = fileService.getTemporaryUniquePath(submissionExportPath, EXPORTED_SUBMISSIONS_DELETION_DELAY_IN_MINUTES);
         try {
-            return exportStudentSubmissions(exerciseId, submissionExportOptions, outputDir, new ArrayList<>(), new ArrayList<>());
+            return exportStudentSubmissions(exerciseId, submissionExportOptions, true, outputDir, new ArrayList<>(), new ArrayList<>());
         }
         catch (IOException e) {
             log.error("Failed to export student submissions for exercise {} to {}: {}", exerciseId, outputDir, e);
-            return Optional.empty();
+            return List.of();
         }
     }
 
     /**
      * Exports student submissions to a zip file for an exercise.
-     *
+     * <p>
      * The outputDir is used to store the zip file and temporary files used for zipping so make
      * sure to delete it if it's no longer used.
      *
@@ -95,13 +103,13 @@ public abstract class SubmissionExportService {
      * @return a reference to the zipped file
      * @throws IOException if an error occurred while zipping
      */
-    public Optional<File> exportStudentSubmissions(Long exerciseId, SubmissionExportOptionsDTO submissionExportOptions, Path outputDir, List<String> exportErrors,
-            List<ArchivalReportEntry> reportData) throws IOException {
+    public List<Path> exportStudentSubmissions(Long exerciseId, SubmissionExportOptionsDTO submissionExportOptions, boolean zipSubmissions, Path outputDir,
+            List<String> exportErrors, List<ArchivalReportEntry> reportData) throws IOException {
 
         Optional<Exercise> exerciseOpt = exerciseRepository.findWithEagerStudentParticipationsStudentAndSubmissionsById(exerciseId);
 
         if (exerciseOpt.isEmpty()) {
-            return Optional.empty();
+            return List.of();
         }
 
         Exercise exercise = exerciseOpt.get();
@@ -133,12 +141,13 @@ public abstract class SubmissionExportService {
         // Sort the student participations by id
         exportedStudentParticipations.sort(Comparator.comparing(DomainObject::getId));
 
-        return createZipFileFromParticipations(exercise, exportedStudentParticipations, enableFilterAfterDueDate, filterLateSubmissionsDate, outputDir, exportErrors, reportData);
+        return exportSubmissionsFromParticipationsOptionallyZipped(exercise, exportedStudentParticipations, enableFilterAfterDueDate, filterLateSubmissionsDate, zipSubmissions,
+                outputDir, exportErrors, reportData);
     }
 
     /**
      * Creates a zip file from a list of participations for an exercise.
-     *
+     * <p>
      * The outputDir is used to store the zip file and temporary files used for zipping so make
      * sure to delete it if it's no longer used.
      *
@@ -149,11 +158,12 @@ public abstract class SubmissionExportService {
      * @param outputDir                directory to store the temporary files in
      * @param exportErrors             a list of errors for submissions that couldn't be exported and are not included in the file
      * @param reportData               a list of all exercises and their statistics
-     * @return the zipped file
+     * @return paths of the exported submissions
      * @throws IOException if an error occurred while zipping
      */
-    private Optional<File> createZipFileFromParticipations(Exercise exercise, List<StudentParticipation> participations, boolean enableFilterAfterDueDate,
-            @Nullable ZonedDateTime lateSubmissionFilter, Path outputDir, List<String> exportErrors, List<ArchivalReportEntry> reportData) throws IOException {
+    private List<Path> exportSubmissionsFromParticipationsOptionallyZipped(Exercise exercise, List<StudentParticipation> participations, boolean enableFilterAfterDueDate,
+            @Nullable ZonedDateTime lateSubmissionFilter, boolean zipSubmissions, Path outputDir, List<String> exportErrors, List<ArchivalReportEntry> reportData)
+            throws IOException {
 
         Course course = exercise.getCourseViaExerciseGroupOrCourseMember();
 
@@ -170,7 +180,7 @@ public abstract class SubmissionExportService {
         if (!submissionFolder.exists() && !submissionFolder.mkdirs()) {
             log.error("Couldn't create dir: {}", submissionFolder);
             exportErrors.add("Cannot create directory: " + submissionFolder.toPath());
-            return Optional.empty();
+            return List.of();
         }
 
         // Create counter for log entry
@@ -207,19 +217,23 @@ public abstract class SubmissionExportService {
                 new ArchivalReportEntry(exercise, FileService.sanitizeFilename(exercise.getTitle()), participations.size(), submissionFilePaths.size(), skippedEntries.intValue()));
 
         if (submissionFilePaths.isEmpty()) {
-            return Optional.empty();
+            return List.of();
         }
 
         // zip stores submissions
-        try {
-            zipFileService.createZipFile(zipFilePath, submissionFilePaths, submissionsFolderPath);
+        if (zipSubmissions) {
+            try {
+                zipFileService.createZipFile(zipFilePath, submissionFilePaths, submissionsFolderPath);
+                return List.of(zipFilePath);
+            }
+            finally {
+                log.debug("Delete all temporary files");
+                fileService.deleteFiles(submissionFilePaths);
+            }
         }
-        finally {
-            log.debug("Delete all temporary files");
-            fileService.deleteFiles(submissionFilePaths);
+        else {
+            return submissionFilePaths;
         }
-
-        return Optional.of(zipFilePath.toFile());
     }
 
     /**
